LPD-16652 Rename the component
