+++ modules/apps/knowledge-base/knowledge-base-web/src/main/resources/META-INF/resources/admin/common/view_kb_article.jsp
@@ -282,7 +282,7 @@ String kbArticleSuccessMessage = GetterUtil.getString(MultiSessionMessages.get(r
 </c:if>
 
 <div>
-	<react:component module="{LockedArticleModal} from knowledge-base-web" props='<%=
+	<react:component module="{LockedKBArticleModal} from knowledge-base-web" props='<%=
 			HashMapBuilder.<String, Object>put(
 				"open", MultiSessionErrors.contains(liferayPortletRequest, DuplicateLockException.class.getName())
 			).build()
