+++ modules/apps/knowledge-base/knowledge-base-web/src/main/resources/META-INF/resources/admin/view.jsp
@@ -162,7 +162,7 @@ String kbArticleSuccessMessage = GetterUtil.getString(MultiSessionMessages.get(r
 
 <div>
 	<react:component
-		module="{LockedArticleModal} from knowledge-base-web"
+		module="{LockedKBArticleModal} from knowledge-base-web"
 		props='<%=
 			HashMapBuilder.<String, Object>put(
 				"open", MultiSessionErrors.contains(liferayPortletRequest, DuplicateLockException.class.getName())
