+++ modules/apps/knowledge-base/knowledge-base-web/src/main/resources/META-INF/resources/js/admin/components/LockedKBArticleModal.js
